% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pallete.R
\docType{data}
\name{Pallete}
\alias{Pallete}
\title{set color pallete
importFrom R6 R6Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Pallete
}
\description{
set color pallete
importFrom R6 R6Class
}
\examples{
\dontrun{
pallete <- Pallete$new("material_design")
print(pallete)
mdc <- pallete$colors

mdc$blue
}
}
\keyword{datasets}

